pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tony_tesla"
version = "0.1.0"
description = "A short description of your project"
authors = [
    { name="Your Name", email="your.email@example.com" }
]
license = { text = "MIT" }
readme = "README.md"
dependencies = [
    "icecream",
    "rich",
    "pydantic",
    "loguru",
    "langchain",
    "httpx",
    "fastapi",
    "azure-cosmos",
    "typer",
    "requests",
    "python-dateutil",
    "pathlib",
    "onebusaway",
    "modal",
    "textual>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio", 
    "pytest-cov",
    "textual-dev",
    "pytest-timeout"
]

[tool.setuptools]
py-modules = ["tony", "storage", "bus", "calls", "cache", "models"]

[project.scripts]
storage = "storage:app"
tony = "tony:app"
bus = "bus:app"
calls = "calls:app"

[tool.pytest.ini_options]
testpaths = ["test_*.py"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
timeout = 300  # 5 minutes timeout for tests
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]

[tool.coverage.run]
source = ["."]
omit = [
    "tests/*",
    "setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "pass",
]
